small refactor
removed some unused files
formatted files
updated tests
