--- hnclient-v2/src/main/java/ca/bc/gov/hlth/hnclientv2/tmp/input/done/E45pvc2.txt
@@ -1,5 +1,0 @@
-MSH|^~\&|HR|BC00000098|RAIGET-DOC-SUM|BC0003000|19991004103039|lharris|E45|19980915000015|D|2.3
-HDR|||TRAININGAdmin
-SFT|1.0||testorg^^orgid^^^MOH|1.0|barebones||
-QPD|E45^^HNET0003|1|^^00000001^^^CANBC^XX^MOH|^^00000001^^^CANBC^XX^MOH|^^00000754^^^CANBC^XX^MOH|9082583512^^^CANBC^JHN^MOH||19770805||||||20020301||PVC^^HNET9909||
-RCP|I|
