--- hnclient-v2/src/main/java/ca/bc/gov/hlth/hnclientv2/tmp/output/E45pvc2.txt
@@ -1,1 +1,0 @@
-TVNIfF5+XCZ8SFJ8QkMwMDAwMDA5OHxSQUlHRVQtRE9DLVNVTXxCQzAwMDMwMDB8MTk5OTEwMDQxMDMwMzl8bGhhcnJpc3xFNDV8MTk5ODA5MTUwMDAwMTV8RHwyLjMNCkhEUnx8fFRSQUlOSU5HQWRtaW4NClNGVHwxLjB8fHRlc3RvcmdeXm9yZ2lkXl5eTU9IfDEuMHxiYXJlYm9uZXN8fA0KUVBEfEU0NV5eSE5FVDAwMDN8MXxeXjAwMDAwMDAxXl5eQ0FOQkNeWFheTU9IfF5eMDAwMDAwMDFeXl5DQU5CQ15YWF5NT0h8Xl4wMDAwMDc1NF5eXkNBTkJDXlhYXk1PSHw5MDgyNTgzNTEyXl5eQ0FOQkNeSkhOXk1PSHx8MTk3NzA4MDV8fHx8fHwyMDAyMDMwMXx8UFZDXl5ITkVUOTkwOXx8DQpSQ1B8SXwNCg==
