+++ hnclient-v2/src/main/resources/application.properties
@@ -30,9 +30,9 @@ client-id = moh_hnclient_dev
 
 # Client authentication method #2: Signed JWT
 ## jks-key-alias should be the same as client id in keycloak
-jks-key-alias = moh-hnclient
-jks-file = C:\\Dev\\git\\HNI\\conf\\moh_hnclient_dev.jks
-cer-file = C:\\Dev\\git\\HNI\\conf\\moh_hnclient_dev.cer
+jks-key-alias = moh_hnclient_dev
+jks-file = C:\\Dev\\Downloads\\moh_hnclient_dev.jks
+cer-file = C:\\Dev\\Downloads\\moh_hnclient_dev.cer
 ## requires the actual keycloak id of the client
 cert-upload-endpoint = https://common-logon-dev.hlth.gov.bc.ca/auth/admin/realms/v2_pos/clients/769006a4-b981-4e33-8ef4-a1b0591d0375/certificates/jwt.credential/upload-certificate
 ### Specify MOH_HNCLIENT_KEYSTORE_PASSWORD as an operating system environment variable
